+++ foxes/models/wake_models/wind/turbopark.py
@@ -18,8 +18,8 @@ class TurbOParkWake(GaussianWakeModel):
 
     Attributes
     ----------
-    A: float
-        The wake growth parameter A.
+    k: float
+        The wake growth parameter k.
     sbeta_factor: float
         Factor multiplying sbeta
     c1: float
@@ -36,7 +36,7 @@ class TurbOParkWake(GaussianWakeModel):
     def __init__(
         self,
         superposition,
-        A,
+        k,
         sbeta_factor=0.25,
         c1=1.5,
         c2=0.8,
@@ -49,8 +49,8 @@ class TurbOParkWake(GaussianWakeModel):
         ----------
         superposition: str
             The wind deficit superposition
-        A: float
-            The wake growth parameter A.
+        k: float
+            The wake growth parameter k.
         sbeta_factor: float
             Factor multiplying sbeta
         c1: float
@@ -63,7 +63,7 @@ class TurbOParkWake(GaussianWakeModel):
         """
         super().__init__(superpositions={FV.WS: superposition})
 
-        self.A = A
+        self.k = k
         self.sbeta_factor = sbeta_factor
         self.c1 = c1
         self.c2 = c2
@@ -74,7 +74,7 @@ class TurbOParkWake(GaussianWakeModel):
             self.induction if isinstance(self.induction, str) else self.induction.name
         )
         s = f"{type(self).__name__}"
-        s += f"({self.superpositions[FV.WS]}, A={self.A}, induction={iname})"
+        s += f"({self.superpositions[FV.WS]}, k={self.k}, induction={iname})"
         return s
 
     def sub_models(self):
@@ -201,7 +201,7 @@ class TurbOParkWake(GaussianWakeModel):
             # calculate sigma (eqn 4)
             sigma = D * (
                 epsilon
-                + self.A
+                + self.k
                 * ati
                 / beta
                 * (
@@ -246,8 +246,8 @@ class TurbOParkWakeIX(GaussianWakeModel):
     ----------
     dx: float
         The step size of the integral
-    A: float
-        The wake growth parameter A.
+    k: float
+        The wake growth parameter k.
     sbeta_factor: float
         Factor multiplying sbeta
     ti_var:  str
@@ -267,7 +267,7 @@ class TurbOParkWakeIX(GaussianWakeModel):
         self,
         superposition,
         dx,
-        A,
+        k,
         sbeta_factor=0.25,
         ti_var=FV.TI,
         self_wake=True,
@@ -283,8 +283,8 @@ class TurbOParkWakeIX(GaussianWakeModel):
             The wind deficit superposition
         dx: float
             The step size of the integral
-        A: float, optional
-            The wake growth parameter A.
+        k: float, optional
+            The wake growth parameter k.
         sbeta_factor: float
             Factor multiplying sbeta
         ti_var:  str
@@ -300,7 +300,7 @@ class TurbOParkWakeIX(GaussianWakeModel):
         super().__init__(superpositions={FV.WS: superposition})
 
         self.dx = dx
-        self.A = A
+        self.k = k
         self.sbeta_factor = sbeta_factor
         self.ti_var = ti_var
         self.ipars = ipars
@@ -313,7 +313,7 @@ class TurbOParkWakeIX(GaussianWakeModel):
             self.induction if isinstance(self.induction, str) else self.induction.name
         )
         s = f"{type(self).__name__}({self.superpositions[FV.WS]}"
-        s += f", ti={self.ti_var}, dx={self.dx}, A={self.A}, induction={iname})"
+        s += f", ti={self.ti_var}, dx={self.dx}, k={self.k}, induction={iname})"
         return s
 
     def sub_models(self):
@@ -473,7 +473,7 @@ class TurbOParkWakeIX(GaussianWakeModel):
             )[:, :, 0]
 
             # calculate sigma (eqn 1, plus epsilon from eqn 4 for x = 0)
-            sigma = D * epsilon + self.A * ti_ix[st_sel]
+            sigma = D * epsilon + self.k * ti_ix[st_sel]
             del x, epsilon
 
             # calculate amplitude, same as in Bastankhah model (eqn 7)